requirements.txt
gunicorn>=22.0.0
Flask==2.3.3
functions-framework==3.9.1
Jinja2>=3.1.0
requests>=2.31.0
google-cloud-firestore>=2.11.0
